HDDS-3601. Refactor TestOzoneManagerHA.java into multiple tests to avoid frequent timeout issues
Closes #930
